requirements.txt
megatron-core>=0.15
modelscope
peft>=0.11,<0.20
safetensors
tqdm
transformers>=4.33,<5.7.0
